[LibOS] Move dl-machine-x86_64.h to arch-specific dir
Move the shim's dl-machine-x86_64.h to include/arch/x86_64/shim_dl-machine.h.
Rename the file to avoid name clashes with Pal's dl-machine.h.
